refactor, code -> annotation def files
