+++ src/neuroglancer/annotation/line_strip.ts
@@ -18,12 +18,24 @@
  * @file Support for rendering line strip annotations.
  */
 
-import {AnnotationType, LineStrip} from 'neuroglancer/annotation';
+import {Annotation, AnnotationReference, AnnotationType, Collection, LineStrip} from 'neuroglancer/annotation';
+import {TwoStepAnnotationTool} from 'neuroglancer/annotation/annotation';
+import {MultiStepAnnotationTool} from 'neuroglancer/annotation/collection';
+import {AnnotationLayerState} from 'neuroglancer/annotation/frontend';
+import {PlaceLineTool} from 'neuroglancer/annotation/line';
+import {PlacePointTool} from 'neuroglancer/annotation/point';
 import {AnnotationRenderContext, AnnotationRenderHelper, registerAnnotationTypeRenderHandler} from 'neuroglancer/annotation/type_handler';
+import {MouseSelectionState} from 'neuroglancer/layer';
+import {StatusMessage} from 'neuroglancer/status';
+import {UserLayerWithAnnotations} from 'neuroglancer/ui/annotations';
+import {registerTool} from 'neuroglancer/ui/tool';
 import {mat4, vec3} from 'neuroglancer/util/geom';
 import {CircleShader} from 'neuroglancer/webgl/circles';
 import {emitterDependentShaderGetter, ShaderBuilder} from 'neuroglancer/webgl/shader';
-// TODO: Ideally, Collection annotations should not be rendered at all, if possible remove drawing code
+// TODO: Collection annotations should not be rendered at all, if possible remove drawing code
+
+const ANNOTATE_LINE_STRIP_TOOL_ID = 'annotateLineStrip';
+
 class RenderHelper extends AnnotationRenderHelper {
   private circleShader = this.registerDisposer(new CircleShader(this.gl));
   private shaderGetter = emitterDependentShaderGetter(
@@ -94,3 +106,99 @@ registerAnnotationTypeRenderHandler(AnnotationType.LINE_STRIP, {
     return annotation;
   }
 });
+
+export class PlaceLineStripTool extends MultiStepAnnotationTool {
+  annotationType: AnnotationType.LINE_STRIP;
+  toolset = PlaceLineTool;
+  looped = false;
+  initMouseState: MouseSelectionState;
+  initPos: any;
+  constructor(public layer: UserLayerWithAnnotations, options: any) {
+    super(layer, options);
+    this.childTool = new this.toolset(layer, options);
+    this.toolbox = options.toolbox;
+    if (this.toolbox.querySelector('.neuroglancer-linestrip-looped')) {
+      this.looped = true;
+    }
+  }
+
+  getInitialAnnotation(mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState):
+      Annotation {
+    const result = <LineStrip>super.getInitialAnnotation(mouseState, annotationLayer);
+    result.connected = true;
+    result.looped = this.looped;
+    result.type = this.annotationType;
+    return result;
+  }
+
+  initChildAnnotation() {
+    if (!(this.childTool instanceof PlacePointTool)) {
+      const inProgressAnnotation = (<TwoStepAnnotationTool>this.childTool).inProgressAnnotation;
+      // Child should not be a collection (for now)
+      const child = inProgressAnnotation!.reference;
+      const disposer = inProgressAnnotation!.disposer;
+      inProgressAnnotation!.disposer = () => {
+        // for every child annotation, redefine its disposer to call the parent to push to the
+        // entry list
+        disposer();
+      };
+      return child;
+    }
+    return;
+  }
+
+  appendNewChildAnnotation(oldAnnotationRef: AnnotationReference, mouseState: MouseSelectionState):
+      Annotation {
+    const oldAnnotation = <Collection>oldAnnotationRef!.value!;
+    this.childTool = new this.toolset(this.layer, {});
+    this.childTool.trigger(mouseState, oldAnnotationRef);
+
+    let last = this.initChildAnnotation();
+    return {...oldAnnotation, last};
+  }
+
+  trigger(mouseState: MouseSelectionState, parentRef?: AnnotationReference) {
+    if (mouseState.active) {
+      if (this.inProgressAnnotation === undefined) {
+        this.initMouseState = <MouseSelectionState>{...mouseState};
+        this.initPos = mouseState.position.slice();
+        super.trigger(mouseState, parentRef);
+      } else {
+        super.trigger(mouseState, parentRef);
+        // Start new annotation automatically
+        this.appendNewChildAnnotation(this.inProgressAnnotation.reference!, mouseState);
+      }
+    }
+  }
+
+  complete() {
+    if (this.inProgressAnnotation) {
+      const innerEntries = (<LineStrip>this.inProgressAnnotation!.reference.value!).entries;
+      if (innerEntries.length > 1) {
+        if (this.looped) {
+          const fakeMouse = <MouseSelectionState>{...this.initMouseState, position: this.initPos};
+          (<LineStrip>this.inProgressAnnotation!.reference.value!).looped = true;
+          this.childTool!.trigger(fakeMouse, this.inProgressAnnotation!.reference);
+        }
+        super.complete();
+      } else {
+        // for LineStrip, a second annotation is created automatically after the first is done
+        // if entries.length is less than 2, the first annotation is not confirmed
+        StatusMessage.showTemporaryMessage(`No annotation has been made.`, 3000);
+      }
+    }
+  }
+
+  get description() {
+    return `annotate line strip ${this.looped ? '(loop)' : ''}`;
+  }
+
+  toJSON() {
+    return ANNOTATE_LINE_STRIP_TOOL_ID;
+  }
+}
+PlaceLineStripTool.prototype.annotationType = AnnotationType.LINE_STRIP;
+
+registerTool(
+    ANNOTATE_LINE_STRIP_TOOL_ID,
+    (layer, options) => new PlaceLineStripTool(<UserLayerWithAnnotations>layer, options));
