+++ src/neuroglancer/ui/annotations.ts
@@ -22,22 +22,28 @@ import './annotations.css';
 
 import debounce from 'lodash/debounce';
 import {Annotation, AnnotationReference, AnnotationSource, AnnotationTag, AnnotationType, AxisAlignedBoundingBox, Collection, Ellipsoid, getAnnotationTypeHandler, Line, LineStrip, LocalAnnotationSource, makeAnnotationId, Point} from 'neuroglancer/annotation';
+import {PlaceAnnotationTool} from 'neuroglancer/annotation/annotation';
+import {PlaceBoundingBoxTool} from 'neuroglancer/annotation/bounding_box';
+import {MultiStepAnnotationTool} from 'neuroglancer/annotation/collection';
+import {PlaceSphereTool} from 'neuroglancer/annotation/ellipsoid';
 import {AnnotationLayer, AnnotationLayerState, PerspectiveViewAnnotationLayer, SliceViewAnnotationLayer} from 'neuroglancer/annotation/frontend';
 import {DataFetchSliceViewRenderLayer, MultiscaleAnnotationSource} from 'neuroglancer/annotation/frontend_source';
+import {PlaceLineTool} from 'neuroglancer/annotation/line';
+import {PlaceLineStripTool} from 'neuroglancer/annotation/line_strip';
+import {PlacePointTool} from 'neuroglancer/annotation/point';
 import {setAnnotationHoverStateFromMouseState} from 'neuroglancer/annotation/selection';
-import {MouseSelectionState, UserLayer} from 'neuroglancer/layer';
+import {UserLayer} from 'neuroglancer/layer';
 import {VoxelSize} from 'neuroglancer/navigation_state';
 import {SegmentationDisplayState} from 'neuroglancer/segmentation_display_state/frontend';
 import {StatusMessage} from 'neuroglancer/status';
 import {TrackableAlphaValue, trackableAlphaValue} from 'neuroglancer/trackable_alpha';
 import {TrackableBoolean, TrackableBooleanCheckbox} from 'neuroglancer/trackable_boolean';
 import {registerNested, TrackableValueInterface, WatchableRefCounted, WatchableValue} from 'neuroglancer/trackable_value';
-import {registerTool, Tool} from 'neuroglancer/ui/tool';
 import {TrackableRGB} from 'neuroglancer/util/color';
 import {Borrowed, Owned, RefCounted} from 'neuroglancer/util/disposable';
 import {removeChildren, removeFromParent} from 'neuroglancer/util/dom';
-import {mat3, mat3FromMat4, mat4, transformVectorByMat4, vec3} from 'neuroglancer/util/geom';
-import {verifyObject, verifyObjectProperty, verifyOptionalInt, verifyOptionalString, verifyString} from 'neuroglancer/util/json';
+import {mat4, transformVectorByMat4, vec3} from 'neuroglancer/util/geom';
+import {verifyObject, verifyObjectProperty, verifyOptionalInt, verifyString} from 'neuroglancer/util/json';
 import {NullarySignal} from 'neuroglancer/util/signal';
 import {formatBoundingBoxVolume, formatIntegerBounds, formatIntegerPoint, formatLength} from 'neuroglancer/util/spatial_units';
 import {Uint64} from 'neuroglancer/util/uint64';
@@ -58,9 +64,6 @@ type AnnotationIdAndPart = {
   multiple?: Set<string>,
   ungroupable?: boolean
 };
-
-type MultiTool = typeof PlacePointTool|typeof PlaceBoundingBoxTool|typeof PlaceLineTool|
-    typeof PlaceSphereTool|typeof PlaceLineStripTool|typeof MultiStepAnnotationTool;
 
 export class AnnotationSegmentListWidget extends RefCounted {
   element = document.createElement('div');
@@ -178,7 +181,6 @@ export class AnnotationSegmentListWidget extends RefCounted {
     });
   }
 }
-
 export class SelectedAnnotationState extends RefCounted implements
     TrackableValueInterface<AnnotationIdAndPart|undefined> {
   private value_: AnnotationIdAndPart|undefined;
@@ -1634,552 +1636,6 @@ export class AnnotationTab extends Tab {
   }
 }
 
-function getSelectedAssocatedSegment(annotationLayer: AnnotationLayerState) {
-  let segments: Uint64[]|undefined;
-  const segmentationState = annotationLayer.segmentationState.value;
-  if (segmentationState != null) {
-    if (segmentationState.segmentSelectionState.hasSelectedSegment) {
-      segments = [segmentationState.segmentSelectionState.selectedSegment.clone()];
-    }
-  }
-  return segments;
-}
-
-abstract class PlaceAnnotationTool extends Tool {
-  temp?: Annotation;
-  group: string;
-  annotationDescription: string|undefined;
-  annotationType: AnnotationType.POINT|AnnotationType.LINE|
-      AnnotationType.AXIS_ALIGNED_BOUNDING_BOX|AnnotationType.ELLIPSOID|
-      AnnotationType.COLLECTION|AnnotationType.LINE_STRIP;
-  constructor(public layer: UserLayerWithAnnotations, options: any) {
-    super();
-    if (layer.annotationLayerState === undefined) {
-      throw new Error(`Invalid layer for annotation tool.`);
-    }
-    this.annotationDescription = verifyObjectProperty(options, 'description', verifyOptionalString);
-  }
-
-  get annotationLayer() {
-    return this.layer.annotationLayerState.value;
-  }
-
-  complete() {
-    StatusMessage.showTemporaryMessage(`Only supported in collection annotations.`, 3000);
-  }
-}
-
-const ANNOTATE_POINT_TOOL_ID = 'annotatePoint';
-const ANNOTATE_LINE_TOOL_ID = 'annotateLine';
-const ANNOTATE_COLLECTION_TOOL_ID = 'annotateCollection';
-const ANNOTATE_LINE_STRIP_TOOL_ID = 'annotateLineStrip';
-const ANNOTATE_BOUNDING_BOX_TOOL_ID = 'annotateBoundingBox';
-const ANNOTATE_ELLIPSOID_TOOL_ID = 'annotateSphere';
-
-export class PlacePointTool extends PlaceAnnotationTool {
-  annotationType: AnnotationType.POINT;
-  constructor(layer: UserLayerWithAnnotations, options: any) {
-    super(layer, options);
-  }
-
-  trigger(mouseState: MouseSelectionState, parentRef?: AnnotationReference) {
-    const {annotationLayer} = this;
-    if (annotationLayer === undefined) {
-      // Not yet ready.
-      return;
-    }
-    if (mouseState.active) {
-      const annotation: Annotation = {
-        id: '',
-        description: '',
-        segments: getSelectedAssocatedSegment(annotationLayer),
-        point:
-            vec3.transformMat4(vec3.create(), mouseState.position, annotationLayer.globalToObject),
-        type: AnnotationType.POINT,
-      };
-      if (parentRef) {
-        annotation.pid = parentRef.id;
-      }
-      const reference = annotationLayer.source.add(annotation, /*commit=*/true);
-      this.layer.selectedAnnotation.value = {id: reference.id};
-      if (parentRef) {
-        const parent = (<Collection>parentRef.value!);
-        parent.entries.push(reference.id);
-        if (parent.segments && annotation.segments) {
-          parent.segments = [...parent.segments!, ...annotation.segments!];
-        }
-      }
-      reference.dispose();
-    }
-  }
-
-  get description() {
-    return `annotate point`;
-  }
-
-  toJSON() {
-    return ANNOTATE_POINT_TOOL_ID;
-  }
-}
-
-function getMousePositionInAnnotationCoordinates(
-    mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState) {
-  return vec3.transformMat4(vec3.create(), mouseState.position, annotationLayer.globalToObject);
-}
-
-abstract class TwoStepAnnotationTool extends PlaceAnnotationTool {
-  inProgressAnnotation:
-      {annotationLayer: AnnotationLayerState, reference: AnnotationReference, disposer: () => void}|
-      undefined;
-
-  abstract getInitialAnnotation(
-      mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState): Annotation;
-  abstract getUpdatedAnnotation(
-      oldAnnotation: Annotation, mouseState: MouseSelectionState,
-      annotationLayer: AnnotationLayerState): Annotation;
-
-  trigger(mouseState: MouseSelectionState, parentRef?: AnnotationReference) {
-    const {annotationLayer} = this;
-    if (annotationLayer === undefined) {
-      // Not yet ready.
-      return;
-    }
-    if (mouseState.active) {
-      const updatePointB = () => {
-        const state = this.inProgressAnnotation!;
-        const reference = state.reference;
-        const newAnnotation =
-            this.getUpdatedAnnotation(reference.value!, mouseState, annotationLayer);
-        state.annotationLayer.source.update(reference, newAnnotation);
-        this.layer.selectedAnnotation.value = {id: reference.id};
-      };
-
-      if (this.inProgressAnnotation === undefined) {
-        const annotation = this.getInitialAnnotation(mouseState, annotationLayer);
-        if (parentRef) {
-          annotation.pid = parentRef.id;
-        }
-        const reference = annotationLayer.source.add(annotation, /*commit=*/false);
-        if (parentRef) {
-          const parent = (<Collection>parentRef.value!);
-          parent.entries.push(reference.id);
-          if (parent.segments && annotation.segments) {
-            parent.segments = [...parent.segments!, ...annotation.segments!];
-          }
-        }
-        this.layer.selectedAnnotation.value = {id: reference.id};
-        const disposer = () => {
-          mouseDisposer();
-          reference.dispose();
-        };
-        this.inProgressAnnotation = {
-          annotationLayer,
-          reference,
-          disposer,
-        };
-        const mouseDisposer = mouseState.changed.add(updatePointB);
-      } else {
-        updatePointB();
-        if (this.inProgressAnnotation) {
-          this.inProgressAnnotation.annotationLayer.source.commit(
-              this.inProgressAnnotation.reference);
-          this.inProgressAnnotation.disposer();
-          this.inProgressAnnotation = undefined;
-          this.layer.selectedAnnotation.changed.dispatch();
-        }
-      }
-    }
-  }
-
-  disposed() {
-    this.deactivate();
-    super.disposed();
-  }
-
-  deactivate() {
-    if (this.inProgressAnnotation !== undefined) {
-      this.inProgressAnnotation.annotationLayer.source.delete(this.inProgressAnnotation.reference);
-      this.inProgressAnnotation.disposer();
-      this.inProgressAnnotation = undefined;
-    }
-  }
-}
-
-abstract class PlaceTwoCornerAnnotationTool extends TwoStepAnnotationTool {
-  annotationType: AnnotationType.LINE|AnnotationType.AXIS_ALIGNED_BOUNDING_BOX;
-
-  getInitialAnnotation(mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState):
-      Annotation {
-    const point = getMousePositionInAnnotationCoordinates(mouseState, annotationLayer);
-    return <AxisAlignedBoundingBox|Line>{
-      id: '',
-      type: this.annotationType,
-      description: '',
-      pointA: point,
-      pointB: point,
-    };
-  }
-
-  getUpdatedAnnotation(
-      oldAnnotation: AxisAlignedBoundingBox|Line, mouseState: MouseSelectionState,
-      annotationLayer: AnnotationLayerState): Annotation {
-    const point = getMousePositionInAnnotationCoordinates(mouseState, annotationLayer);
-    return {...oldAnnotation, pointB: point};
-  }
-}
-
-export class MultiStepAnnotationTool extends PlaceAnnotationTool {
-  inProgressAnnotation:
-      {annotationLayer: AnnotationLayerState, reference: AnnotationReference, disposer: () => void}|
-      undefined;
-  annotationType: AnnotationType.COLLECTION|AnnotationType.LINE_STRIP;
-  toolset: MultiTool;
-  toolbox: HTMLDivElement;
-  childTool: PlacePointTool|PlaceBoundingBoxTool|PlaceLineTool|PlaceSphereTool|PlaceLineStripTool|
-      undefined;
-  constructor(public layer: UserLayerWithAnnotations, options: any) {
-    super(layer, options);
-    this.toolbox = options.toolbox;
-  }
-
-  getInitialAnnotation(mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState):
-      Annotation {
-    const coll = <Collection>{
-      id: '',
-      type: this.annotationType,
-      description: '',
-      entries: [],
-      segments: [],
-      connected: false,
-      source:
-          vec3.transformMat4(vec3.create(), mouseState.position, annotationLayer.globalToObject),
-      entry: () => {},
-      cVis: new TrackableBoolean(true, true)
-    };
-    coll.entry = (index: number) =>
-        (<LocalAnnotationSource>annotationLayer.source).get(coll.entries[index]);
-    return coll;
-  }
-  complete() {
-    if ((this.inProgressAnnotation && this.childTool) ||
-        (this.inProgressAnnotation && this.inProgressAnnotation!.reference.value! &&
-         (<Collection>this.inProgressAnnotation!.reference.value!).entries.length)) {
-      const childInProgress = this.childTool ?
-          (<MultiStepAnnotationTool|TwoStepAnnotationTool>this.childTool!).inProgressAnnotation :
-          void (0);
-      const childCount = (<Collection>this.inProgressAnnotation!.reference.value!).entries.length;
-      if (this.childTool && (<PlaceLineStripTool>this.childTool!).toolset) {
-        if ((<LineStrip>childInProgress!.reference.value!).entries.length > 1) {
-          this.childTool!.complete();
-          this.childTool!.dispose();
-          StatusMessage.showTemporaryMessage(
-              `Child annotation ${childInProgress!.reference.value!.id} complete.`);
-          this.childTool = undefined;
-          this.layer.tool.changed.dispatch();
-          this.layer.selectedAnnotation.changed.dispatch();
-
-          let key = this.toolbox.querySelector('.neuroglancer-child-collection-tool');
-          if (key) {
-            key.classList.remove('neuroglancer-child-collection-tool');
-          }
-        } else {
-          // see line 1960
-          StatusMessage.showTemporaryMessage(`No annotation has been made.`, 3000);
-        }
-      } else if ((!childInProgress && childCount === 1) || childCount > 1) {
-        if (this.childTool) {
-          this.childTool!.dispose();
-        }
-        this.inProgressAnnotation!.annotationLayer.source.commit(
-            this.inProgressAnnotation!.reference);
-        StatusMessage.showTemporaryMessage(
-            `Annotation ${this.inProgressAnnotation!.reference.value!.id} complete.`);
-        this.inProgressAnnotation!.disposer();
-        this.inProgressAnnotation = undefined;
-        this.layer.selectedAnnotation.changed.dispatch();
-      } else {
-        StatusMessage.showTemporaryMessage(`No annotation has been made.`, 3000);
-      }
-    } else {
-      // if child tool has a toolset, its a lineStrip, and we apply the lineStrip test b4
-      // continuing if child tool is a base annotation, it must have at least one complete
-      // annotation an annotation is complete if the child tool has no inProgressAnnotation if
-      // there are more than two annotations in entries, the first one is guaranteed to be
-      // complete
-      StatusMessage.showTemporaryMessage(`No annotation has been made.`, 3000);
-    }
-  }
-
-  trigger(mouseState: MouseSelectionState, parentRef?: AnnotationReference) {
-    const {annotationLayer} = this;
-    if (annotationLayer === undefined) {
-      // Not yet ready.
-      return;
-    }
-    if (!this.childTool) {
-      StatusMessage.showTemporaryMessage(`Collections require a child tool.`);
-      return;
-    }
-    if (mouseState.active) {
-      if (this.inProgressAnnotation === undefined) {
-        const annotation = this.getInitialAnnotation(mouseState, annotationLayer);
-        if (parentRef) {
-          annotation.pid = parentRef.id;
-        }
-        const reference = annotationLayer.source.add(annotation, /*commit=*/false);
-        if (parentRef) {
-          const parent = (<Collection>parentRef.value!);
-          parent.entries.push(reference.id);
-          if (parent.segments && annotation.segments) {
-            parent.segments = [...parent.segments!, ...annotation.segments!];
-          }
-        }
-        this.layer.selectedAnnotation.value = {id: reference.id};
-        this.childTool.trigger(mouseState, /*child=*/reference);
-        const disposer = () => {
-          mouseDisposer();
-          reference.dispose();
-        };
-        this.inProgressAnnotation = {
-          annotationLayer,
-          reference,
-          disposer,
-        };
-        const mouseDisposer = () => {};
-      } else {
-        this.childTool.trigger(mouseState, this.inProgressAnnotation.reference);
-      }
-    }
-  }
-
-  get description() {
-    return `annotate collection: ${
-        this.childTool ? this.childTool.description : 'no child tool selected'}`;
-  }
-
-  toJSON() {
-    return ANNOTATE_COLLECTION_TOOL_ID;
-  }
-
-  dispose() {
-    if (this.childTool) {
-      this.childTool!.dispose();
-    }
-    if (this.inProgressAnnotation) {
-      // completely delete the annotation
-      const annotation_ref = this.inProgressAnnotation.reference;
-      this.annotationLayer!.source.delete(annotation_ref, true);
-      //  childDeleted.dispatch(annotation_ref.value!.id);
-      this.inProgressAnnotation!.disposer();
-    }
-    super.dispose();
-  }
-}
-MultiStepAnnotationTool.prototype.annotationType = AnnotationType.COLLECTION;
-
-export class PlaceLineStripTool extends MultiStepAnnotationTool {
-  annotationType: AnnotationType.LINE_STRIP;
-  toolset = PlaceLineTool;
-  looped = false;
-  initMouseState: MouseSelectionState;
-  initPos: any;
-  constructor(public layer: UserLayerWithAnnotations, options: any) {
-    super(layer, options);
-    this.childTool = new this.toolset(layer, options);
-    this.toolbox = options.toolbox;
-    if (this.toolbox.querySelector('.neuroglancer-linestrip-looped')) {
-      this.looped = true;
-    }
-  }
-
-  getInitialAnnotation(mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState):
-      Annotation {
-    const result = <LineStrip>super.getInitialAnnotation(mouseState, annotationLayer);
-    result.connected = true;
-    result.looped = this.looped;
-    result.type = this.annotationType;
-    return result;
-  }
-
-  initChildAnnotation() {
-    if (!(this.childTool instanceof PlacePointTool)) {
-      const inProgressAnnotation = (<TwoStepAnnotationTool>this.childTool).inProgressAnnotation;
-      // Child should not be a collection (for now)
-      const child = inProgressAnnotation!.reference;
-      const disposer = inProgressAnnotation!.disposer;
-      inProgressAnnotation!.disposer = () => {
-        // for every child annotation, redefine its disposer to call the parent to push to the
-        // entry list
-        disposer();
-      };
-      return child;
-    }
-    return;
-  }
-
-  appendNewChildAnnotation(oldAnnotationRef: AnnotationReference, mouseState: MouseSelectionState):
-      Annotation {
-    const oldAnnotation = <Collection>oldAnnotationRef!.value!;
-    this.childTool = new this.toolset(this.layer, {});
-    this.childTool.trigger(mouseState, oldAnnotationRef);
-
-    let last = this.initChildAnnotation();
-    return {...oldAnnotation, last};
-  }
-
-  trigger(mouseState: MouseSelectionState, parentRef?: AnnotationReference) {
-    if (mouseState.active) {
-      if (this.inProgressAnnotation === undefined) {
-        this.initMouseState = <MouseSelectionState>{...mouseState};
-        this.initPos = mouseState.position.slice();
-        super.trigger(mouseState, parentRef);
-      } else {
-        super.trigger(mouseState, parentRef);
-        // Start new annotation automatically
-        this.appendNewChildAnnotation(this.inProgressAnnotation.reference!, mouseState);
-      }
-    }
-  }
-
-  complete() {
-    if (this.inProgressAnnotation) {
-      const innerEntries = (<LineStrip>this.inProgressAnnotation!.reference.value!).entries;
-      if (innerEntries.length > 1) {
-        if (this.looped) {
-          const fakeMouse = <MouseSelectionState>{...this.initMouseState, position: this.initPos};
-          (<LineStrip>this.inProgressAnnotation!.reference.value!).looped = true;
-          this.childTool!.trigger(fakeMouse, this.inProgressAnnotation!.reference);
-        }
-        super.complete();
-      } else {
-        // for LineStrip, a second annotation is created automatically after the first is done
-        // if entries.length is less than 2, the first annotation is not confirmed
-        StatusMessage.showTemporaryMessage(`No annotation has been made.`, 3000);
-      }
-    }
-  }
-
-  get description() {
-    return `annotate line strip ${this.looped ? '(loop)' : ''}`;
-  }
-
-  toJSON() {
-    return ANNOTATE_LINE_STRIP_TOOL_ID;
-  }
-}
-PlaceLineStripTool.prototype.annotationType = AnnotationType.LINE_STRIP;
-
-export class PlaceBoundingBoxTool extends PlaceTwoCornerAnnotationTool {
-  get description() {
-    return `annotate bounding box`;
-  }
-
-  toJSON() {
-    return ANNOTATE_BOUNDING_BOX_TOOL_ID;
-  }
-}
-PlaceBoundingBoxTool.prototype.annotationType = AnnotationType.AXIS_ALIGNED_BOUNDING_BOX;
-
-export class PlaceLineTool extends PlaceTwoCornerAnnotationTool {
-  get description() {
-    return `annotate line`;
-  }
-
-  getInitialAnnotation(mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState):
-      Annotation {
-    const result = super.getInitialAnnotation(mouseState, annotationLayer);
-    result.segments = getSelectedAssocatedSegment(annotationLayer);
-    return result;
-  }
-
-  getUpdatedAnnotation(
-      oldAnnotation: Line|AxisAlignedBoundingBox, mouseState: MouseSelectionState,
-      annotationLayer: AnnotationLayerState) {
-    const result = super.getUpdatedAnnotation(oldAnnotation, mouseState, annotationLayer);
-    const segments = result.segments;
-    if (segments !== undefined && segments.length > 0) {
-      segments.length = 1;
-    }
-    let newSegments = getSelectedAssocatedSegment(annotationLayer);
-    if (newSegments && segments) {
-      newSegments = newSegments.filter(x => segments.findIndex(y => Uint64.equal(x, y)) === -1);
-    }
-    result.segments = [...(segments || []), ...(newSegments || [])] || undefined;
-    return result;
-  }
-
-  toJSON() {
-    return ANNOTATE_LINE_TOOL_ID;
-  }
-}
-PlaceLineTool.prototype.annotationType = AnnotationType.LINE;
-
-class PlaceSphereTool extends TwoStepAnnotationTool {
-  getInitialAnnotation(mouseState: MouseSelectionState, annotationLayer: AnnotationLayerState):
-      Annotation {
-    const point = getMousePositionInAnnotationCoordinates(mouseState, annotationLayer);
-
-    return <Ellipsoid>{
-      type: AnnotationType.ELLIPSOID,
-      id: '',
-      description: '',
-      segments: getSelectedAssocatedSegment(annotationLayer),
-      center: point,
-      radii: vec3.fromValues(0, 0, 0),
-    };
-  }
-
-  getUpdatedAnnotation(
-      oldAnnotation: Ellipsoid, mouseState: MouseSelectionState,
-      annotationLayer: AnnotationLayerState) {
-    const spatialCenter =
-        vec3.transformMat4(vec3.create(), oldAnnotation.center, annotationLayer.objectToGlobal);
-
-    const radius = vec3.distance(spatialCenter, mouseState.position);
-
-    const tempMatrix = mat3.create();
-    tempMatrix[0] = tempMatrix[4] = tempMatrix[8] = 1 / (radius * radius);
-
-
-    const objectToGlobalLinearTransform =
-        mat3FromMat4(mat3.create(), annotationLayer.objectToGlobal);
-    mat3.multiply(tempMatrix, tempMatrix, objectToGlobalLinearTransform);
-    mat3.transpose(objectToGlobalLinearTransform, objectToGlobalLinearTransform);
-    mat3.multiply(tempMatrix, objectToGlobalLinearTransform, tempMatrix);
-
-    return <Ellipsoid>{
-      ...oldAnnotation,
-      radii: vec3.fromValues(
-          1 / Math.sqrt(tempMatrix[0]), 1 / Math.sqrt(tempMatrix[4]), 1 / Math.sqrt(tempMatrix[8])),
-    };
-  }
-  get description() {
-    return `annotate ellipsoid`;
-  }
-
-  toJSON() {
-    return ANNOTATE_ELLIPSOID_TOOL_ID;
-  }
-}
-
-registerTool(
-    ANNOTATE_POINT_TOOL_ID,
-    (layer, options) => new PlacePointTool(<UserLayerWithAnnotations>layer, options));
-registerTool(
-    ANNOTATE_BOUNDING_BOX_TOOL_ID,
-    (layer, options) => new PlaceBoundingBoxTool(<UserLayerWithAnnotations>layer, options));
-registerTool(
-    ANNOTATE_LINE_TOOL_ID,
-    (layer, options) => new PlaceLineTool(<UserLayerWithAnnotations>layer, options));
-registerTool(
-    ANNOTATE_ELLIPSOID_TOOL_ID,
-    (layer, options) => new PlaceSphereTool(<UserLayerWithAnnotations>layer, options));
-registerTool(
-    ANNOTATE_COLLECTION_TOOL_ID,
-    (layer, options) => new MultiStepAnnotationTool(<UserLayerWithAnnotations>layer, options));
-registerTool(
-    ANNOTATE_LINE_STRIP_TOOL_ID,
-    (layer, options) => new PlaceLineStripTool(<UserLayerWithAnnotations>layer, options));
 
 export interface UserLayerWithAnnotations extends UserLayer {
   annotationLayerState: WatchableRefCounted<AnnotationLayerState>;
